refactor : GameConsole 객체 도메인 패키지로 이동
